Do we need to run setup-sh?
@zhileiz 
Small note: I noticed that the docker documentation on gdocs:
https://docs.google.com/document/d/1A9W1PdZQFWhKyDn-mCjpWZb6D1U-BQ_p/edit
Differs from the README on github
https://github.com/CIS380/docker-env
In that the docker documentation on gdocs does not mention you have to run ./docker-setup.sh.  Indeed, I didn't run docker-setup.sh and it seems like everything worked fine for me.  So not sure if that is actually an accurate step needed in the github readme?
If it is, feel free to disregard this and I'll close out this branch.